auto zoom depending on result amount
