+++ frontend/src/components/home/Map.jsx
@@ -3,21 +3,9 @@ import GoogleMapReact from 'google-map-react';
 import { fitBounds } from 'google-map-react';
 import Marker from "./Marker";
 import "../../styles/css/map.css";
-
+import {markerLoc, boundCoord} from '../../helpers/selector'
 
 const config = {
-  zoomLevel: 9,
-  lat: 45.424721,
-  lng: -75.695000,
-  bounds: {nw: {
-    lat: 44.614721,
-    lng: -117.945000
-    },
-    se: {
-      lat: 33.428335045970396,
-      lng: -117.6217273125
-    }
-  },
   size:{
     width: 640, // Map width in pixels
     height: 380, // Map height in pixels
@@ -25,10 +13,10 @@ const config = {
 }
 
 export default function Map(props) {
-  //const [zoomLevel, setZoomLevel] = useState(config.zoomLevel || 9);
   const [lat, setLat] = useState( 45.424721);
   const [lng, setLng] = useState( -75.695000);
   const [coord, setCoord] = useState([{lat:45.424721, lng:-75.695000}]);
+  //const [multiple, setMultiple] = useState(false);
   const [bounds, setBounds] = useState({nw: {
                                           lat: 45.424721,
                                           lng: -75.695000
@@ -40,67 +28,12 @@ export default function Map(props) {
                                         })
 
   const { mapResults } = props;
-  const {center, zoom} = fitBounds(bounds, config.size);
 
-  const markerLoc = function(resultArr) {
-    let result = [];
-    if (resultArr.length > 0) {
-      
-      for (let obj of resultArr) {
-        let coord = {};
-        if(obj.geography) {
-          coord['lat'] = obj.geography.latitude;
-          coord['lng'] = obj.geography.longitude;
-        }
-      result.push(coord);
-      }
-      return result;
-    } else {
-      return null;
-    }
-  }
-  const boundCoord = function(resultArr) {
-    let result = {};
-    if (resultArr.length > 0) {
-      // find nw marker coord in the objects of markerloc funstion arr
-      // to go furthest north max lat
-      // to go furthest west  min lng
-      resultArr = markerLoc(resultArr)
-      let tempLat = [];
-      let tempLng = [];
-      let tempNWObj = {};
-      let tempSEObj = {};
-      for (let point of resultArr) {
-        tempLat.push(point.lat);
-        tempLng.push(point.lng)
-      }
-      //console.log('tempLat', tempLat)
-      //console.log('tempLng', tempLng)
-
-      tempNWObj['lat'] = tempLat.reduce(function(a, b) {
-        return Math.max(a, b);
-      });
-      //console.log('maxlat', tempObj['lat'])
-    
-      tempNWObj['lng'] = tempLng.reduce(function(a, b) {
-        return Math.min(a, b);
-      });
-
-      result['nw'] =  tempNWObj;
-      //find se marker
-      // to go furthest south min lat
-      // to go furthest west  min lng
-      tempSEObj['lat'] = tempLat.reduce(function(a, b) {
-        return Math.min(a, b);
-      });
-      tempSEObj['lng'] = tempLng.reduce(function(a, b) {
-        return Math.max(a, b);
-      });
-      result['se'] =  tempSEObj;
-      return result;
-    } else {
-      return null;
-    }
+  let {center, zoom} = fitBounds(bounds, config.size);
+  if (mapResults.length === 1) {
+    let coord = markerLoc(mapResults);
+    center = {lat: coord[0].lat, lng: coord[0].lng};
+    zoom = 7;
   }
 
   useEffect(() => { 
@@ -112,10 +45,9 @@ export default function Map(props) {
       setLng(calculatedCoord[0].lng) 
     } 
     let calculatedBoundCoord = boundCoord(mapResults);
-    if(calculatedBoundCoord){
-      console.log("inside useffect", calculatedBoundCoord);
+    if(mapResults.length > 1 && calculatedBoundCoord){
       setBounds(calculatedBoundCoord)
-    }
+    } 
   }, [props]);
 
 
@@ -124,13 +56,13 @@ export default function Map(props) {
       <GoogleMapReact
         className='map'
         bootstrapURLKeys={{key: process.env.REACT_APP_GOOGLE_API}}
-        defaultZoom={9}
+        defaultZoom={5}
         defaultCenter={{
           lat,
           lng
           }}
         center={center}
-        zoom={zoom}
+        zoom={zoom}  
       >
       {/* {marker(results) && marker(results)} */}
       {coord && coord.map((flight, index)=> {
@@ -140,8 +72,7 @@ export default function Map(props) {
         lng={flight.lng}
         />)
       })
-      }
-      
+      } 
       </GoogleMapReact>
     </div>
   );
