+++ frontend/src/components/home/Marker.jsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 const config = {
   markerLongitude: 45.424721,
   markerLatitude: -75.695000,
-  googleMapsMarkerIcon: 'https://developers.google.com/maps/documentation/javascript/examples/full/images/beachflag.png'
+  googleMapsMarkerIcon: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRINp6ut9j1blvcfrkgxQpYDXfnW_LhGhOXcg&usqp=CAU'
 }
 
 export default function Marker(props) {
@@ -15,5 +15,6 @@ export default function Marker(props) {
     src={config.googleMapsMarkerIcon}
     onClick={props.onClick} 
     />
+    // <i className="fas fa-plane"></i>
   )
 }
